Added a waypoint super class and started facings script
Make each individual class a subclass of the waypoint class, like I did
in the facings script. Facing script has the free look code, but not in
the correct format for implementation. waypoint class is currently blank
but will be the code that runs everything. It'll call the
functions/states from the other classes as needed.